+++ app/src/pages/assets/AssetsPage.js
@@ -10,18 +10,33 @@ import WidgetList from '../../components/Widget/WidgetList'
 import qs from 'query-string'
 import './Assets.css'
 
+const getHeadTitle = (type, searchParams) => {
+  switch (type) {
+    case 'all':
+      return 'All Assets'
+    case 'currencies':
+      return 'Currencies'
+    case 'erc20':
+      return 'ERC20 Assets'
+    case 'list':
+      return (qs.parse(searchParams).name || '').split('@')[0].toUpperCase()
+    default:
+      return 'Assets'
+  }
+}
+
 const AssetsPage = props => (
   <div className='page projects-table'>
     <Helmet>
       <title>Assets</title>
       <link rel='canonical' href={`${getOrigin()}/assets`} />
     </Helmet>
-    {qs.parse(props.location.search).feature === 'widgets' && (
+    {props.isBetaModeEnabled && (
       <WidgetList type={props.type} isLoggedIn={props.isLoggedIn} />
     )}
     <div className='page-head page-head-projects'>
       <div className='page-head-projects__left'>
-        <h1>Assets</h1>
+        <h1>{getHeadTitle(props.type, props.location.search)}</h1>
         <HelpPopupAssets />
         {props.type === 'list' &&
           props.location.hash !== '#shared' && <WatchlistShare />}